Remove unnecessary layer. Implement fog of war mesh filter setter
