--- Assets/Scripts/TileVisibility/FogOfWar/FogOfWarMeshTileVisibilityDelegate.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 511843effd044ca49f33fee4c5dd8681
-timeCreated: 1579121814
